Add login shell scripts to set DCONF_PROFILE.
These scripts tell fcmdr-service to write an updated dconf profile
for the current user, and if the D-Bus call is successful sets the
DCONF_PROFILE environment variable.

This lockdown method can be trivially circumvented by a user, but
it's the only mechanism dconf currently offers for establishing a
user-specific profile.