FCmdrProfileSource: Make load_remote() method async.
I'm thinking FCmdrService is going to invoke load_remote() concurrently
for all profile sources, and then collate the results and figure out if
any new or updated profiles need to be applied.

The key point is to only call fcmdr_service_apply_profiles() at most
once per update cycle, so we'll want to know when the load_remote()
operations finish.