Add FCmdrYumBackend.
A new service backend that acts on the "org.baseurl.yum" member of a
profile's settings object.  The JSON member is just the yum repo file
content as a string, which defines one or more yum repositories.  The
service backend simply writes the content verbatim to:

   /etc/yum.repos.d/fleet-commander-$PROFILE_UID.repo

Repo files are deleted automatically when the corresponding profile
is no longer published by the server.