Add FCmdrGoaBackend.
New type of FCmdrServiceBackend to handle GNOME Online Account data
in profiles.  Relies on FCmdrUserResolver to substitute "${username}"
and "${realname}" variables in account details.  It then spawns a
helper utility (fcmdr-update-goa) which sets its UID to match the
logged in user so it can connect to the user's session bus and add
the accounts.

This adds a dependency on goa-1.0 >= 3.15.1.  The bleeding-edge minimum
version is for GOA features added explicitly to support Fleet Commander:

    https://bugzilla.gnome.org/737656
    https://bugzilla.gnome.org/737721